% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCoexpressionNetwork.R
\name{buildCoexpressionNetwork}
\alias{buildCoexpressionNetwork}
\title{Identify the genes that have a positive correlation above a threshold
correlation. Here the threshold correlation was chosen as the median
correlation out of all the positive correlations.}
\usage{
buildCoexpressionNetwork(correlation_matrix)
}
\arguments{
\item{correlation_matrix}{A matrix containing the correlation between all
pairs of genes}
}
\value{
A data frame storing pairs of genes and the correlation between
the genes
}
\description{
Identify the genes that have a positive correlation above a threshold
correlation. Here the threshold correlation was chosen as the median
correlation out of all the positive correlations.
}
\examples{
# Using pbmc_srat.rds file in data folder
srat <- readRds(file = "data/pbmc_srat.rds")
expr_matrix <- getExpressionMatrix(srat = srat)
corr_matrix <- getCorrelationMatrix(expression_matrix = expr_matrix)
edge_list <- buildCoexpressionNetwork(correlation_matrix = corr_matrix)
    
}
