% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDifferentialCoexpression.R
\name{checkForIncorrectInput}
\alias{checkForIncorrectInput}
\title{Check if the expression matrix and coexpression network are valid input}
\usage{
checkForIncorrectInput(exprMatrix, gene1, gene2)
}
\arguments{
\item{exprMatrix}{an expression matrix}

\item{gene1}{a gene name}

\item{gene2}{a different gene name}
}
\value{
no return value but throw an error if applicable
}
\description{
Check if the expression matrix and coexpression network are valid input
}
