% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDifferentialCoexpression.R
\name{checkGeneCoexpression}
\alias{checkGeneCoexpression}
\title{Given two genes, and a a dataframe that contains pairs of genes that are
coexpressed in two networks, check if the two genes are in the dataframe.
If the genes are not in the dataframe, this pair of genes is not
coexpressed in both networks.}
\usage{
checkGeneCoexpression(mergedList, gene1, gene2)
}
\arguments{
\item{mergedList}{a dataframe containing pairs of genes that are coexpressed
in two networks, columns for the correlation in each network, and the
log2 fold change of the correlation}

\item{gene1}{a gene name}

\item{gene2}{another gene name}
}
\value{
the row in mergedList containing gene1 and gene2 if they are
coexpressed in both networks
}
\description{
Given two genes, and a a dataframe that contains pairs of genes that are
coexpressed in two networks, check if the two genes are in the dataframe.
If the genes are not in the dataframe, this pair of genes is not
coexpressed in both networks.
}
