% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{correlationMatrixDC}
\alias{correlationMatrixDC}
\title{Another example correlation matrix that was generated using the same PBMC
data. To generate the data, we ran prepareData using the example
data in the package, getExpressionMatrix with the cell type
set to "DC", and finally ran getCorrelationMatrix with this expression
matrix (other arguments were left as the default). This object was included
to make the examples in the function descriptions run faster.}
\format{
A dataframe with 646 observations and 646 variables
\describe{
each observation and variable are genes, and each cell stores the
correlation between the genes
}
}
\usage{
correlationMatrixDC
}
\description{
Another example correlation matrix that was generated using the same PBMC
data. To generate the data, we ran prepareData using the example
data in the package, getExpressionMatrix with the cell type
set to "DC", and finally ran getCorrelationMatrix with this expression
matrix (other arguments were left as the default). This object was included
to make the examples in the function descriptions run faster.
}
\examples{
\dontrun{
correlationMatrixDC
}
}
\keyword{datasets}
