% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expressionMatrixDC}
\alias{expressionMatrixDC}
\title{Another example expression matrix that was generated using the same PBMC
data. To generate the data, we ran prepareData using the example
data in the package, and then ran getExpressionMatrix with the cell type
set to "DC". This object was included to make the examples in the
function descriptions run faster.}
\format{
A dataframe with 2000 observations and 344 variables
\describe{
each observation is a gene, and each variable is a cell barcode
}
}
\usage{
expressionMatrixDC
}
\description{
Another example expression matrix that was generated using the same PBMC
data. To generate the data, we ran prepareData using the example
data in the package, and then ran getExpressionMatrix with the cell type
set to "DC". This object was included to make the examples in the
function descriptions run faster.
}
\examples{
\dontrun{
expressionMatrixDC
}
}
\keyword{datasets}
