% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoexpressionNetwork.R
\name{getCoexpressionNetwork}
\alias{getCoexpressionNetwork}
\title{Identify the genes that have a positive correlation strictly higher than a
threshold correlation.}
\usage{
getCoexpressionNetwork(correlationMatrix, thresholdCorrelation = 0.8)
}
\arguments{
\item{correlationMatrix}{A matrix or data framer where each row and column
are genes, and each cell contains the correlation between the pairs of
genes.}

\item{thresholdCorrelation}{A percentile used to select only the genes with
a high correlation. This value must be in the range \link{0, 1}. The default
is 0.80, so only gene pairs that have a correlation higher than the
correlation at the 80th percentile will be used.}
}
\value{
A data frame storing pairs of genes and the correlation between
the genes that have a positive correlation above the threshold
}
\description{
Identify the genes that have a positive correlation strictly higher than a
threshold correlation.
}
\examples{
# Using an example correlation matrix that was generated using the pbmc data
# that is available in this package. This matrix only includes platelet
# cells.
corrMatrixPath <- system.file("extdata", 
                              "correlationMatrixPlatelet.csv", 
                               package = "DiffCoExpr")
corrMatrix <- read.csv(corrMatrixPath)
# Set the row names and remove the column containing gene names
rownames(corrMatrix) <- corrMatrix[, 1]
corrMatrix <- corrMatrix[, -1]

coexprNetwork <- getCoexpressionNetwork(correlationMatrix = corrMatrix)
coexprNetwork

}
