% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrelationMatrix.R
\name{getCorrelationMatrix}
\alias{getCorrelationMatrix}
\title{Create a correlation matrix. Lowly expressed genes, and genes will too
little variation across cells are filtered out. In the matrix, each
row and column are genes, and each cell contains the correlation between the
pair of genes.}
\usage{
getCorrelationMatrix(expressionMatrix, minCellCount = 5, minGeneCount = 5)
}
\arguments{
\item{expressionMatrix}{A matrix containing the expression levels of genes
per cell. The rows are genes and the columns are cells}

\item{minCellCount}{A minimum number of cells that should still be present
in the data after filtering. If the number of cells after filtering is
less than this value, stop running. The default is set to 5.}

\item{minGeneCount}{A minimum number of genes that should still be present
in the data after filtering. If the number of genes after filtering is
less than this value, stop running. The default is set to 5.}
}
\value{
A matrix where the rows and columns are genes, and the cell contains
the correlation between the two genes
}
\description{
Create a correlation matrix. Lowly expressed genes, and genes will too
little variation across cells are filtered out. In the matrix, each
row and column are genes, and each cell contains the correlation between the
pair of genes.
}
\examples{
# Using an example expression matrix that was generated using the pbmc data
# that is available in this package. This matrix only includes platelet
# cells.
exprMatrixPath <- system.file("extdata", 
                              "expressionMatrix.csv", 
                               package = "DiffCoExpr")
exprMatrix <- read.csv(exprMatrixPath)
# Set the row names and remove the column containing gene names
rownames(exprMatrix) <- exprMatrix[, 1]
exprMatrix <- exprMatrix[, -1]

corrMatrix <- getCorrelationMatrix(expressionMatrix = exprMatrix)
corrMatrix

}
\references{
Lemoine G., Scott-Boyer M., Ambroise B., Perin O., Droit A. (2021) GWENA:
gene co-expression networks analysis and extended modules characterization
in a single Bioconductor package. BMC Bioinformatics 22, 267.
\href{https://doi.org/10.1186/s12859-021-04179-4}{Link}.
}
