% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentialCoexpression.R
\name{getDifferentialCoexpression}
\alias{getDifferentialCoexpression}
\title{Given two coexpression networks, identify the pairs that are coexpressed
in both networks. Return a dataframe of all the genes that are coexpressed
in both networks, and quantify differential coexpression using log fold
change.}
\usage{
getDifferentialCoexpression(networkA, networkB, thresholdLogFC = 0.5)
}
\arguments{
\item{networkA}{An edge list for a coexpression network. Contains the
correlation coefficient between pairs of genes.}

\item{networkB}{An edge list for a different co-expression network.
Contains the correlation coefficient between pairs of genes.}

\item{thresholdLogFC}{A minimum log2 fold change to be used for filtering
out gene pairs. Default is 0.5}
}
\value{
a dataframe where each row contains a pair of genes and the
correlation of the genes in each network
}
\description{
Given two coexpression networks, identify the pairs that are coexpressed
in both networks. Return a dataframe of all the genes that are coexpressed
in both networks, and quantify differential coexpression using log fold
change.
}
\examples{
# Using an example data that was generated using the pbmc data that is
# available in this package. 

# Condition A: Platelet cells
coexprNetPlateletPath <- system.file("extdata", 
                                     "coexpressionNetworkPlatelet.csv", 
                                     package = "DiffCoExpr")
coexprNetPlatelet <- read.csv(coexprNetPlateletPath)
# Delete the first column 
coexprNetPlatelet <- coexprNetPlatelet[, -1]

# Condition B: DC Cells
coexprNetDCPath <- system.file("extdata", 
                               "coexpressionNetworkDC.csv", 
                               package = "DiffCoExpr")
coexprNetDC <- read.csv(coexprNetDCPath)
# Delete the first column 
coexprNetDC <- coexprNetDC[, -1]

diffCoexpressed <- getDifferentialCoexpression(networkA = coexprNetPlatelet, 
                                               networkB = coexprNetDC) 
diffCoexpressed

}
