% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpressionMatrix.R
\name{getExpressionMatrix}
\alias{getExpressionMatrix}
\title{Create (and visualize) the expression matrix for single-cell RNA-seq data}
\usage{
getExpressionMatrix(srat, cell_type = NULL)
}
\arguments{
\item{srat}{A Seurat object containing normalized counts from a scRNA_seq
experiment.}

\item{cell_type}{A cell type that should be used to filter the expression
matrix so all cells are of that cell type. Can also be a cluster if
cell types were not used. Assumes cell type/cluster are in the Seurat
object. Default is set to NULL, so all cells of all types are included
in the expression matrix}
}
\value{
A dense matrix where the rows are genes and the columns are cell
barcodes. Each cell contains the log normalized counts of that gene in
that cell
}
\description{
Create (and visualize) the expression matrix for single-cell RNA-seq data
}
\examples{
# Using saved Seurat object in the Data folder
srat <- readRds(file = "data/pbmc_srat.rds")
expr_matrix <- get_expression_matrix(srat = srat)
expr_matrix

}
\references{
Butler, A. (2015). Seurat: Tools for Single Cell Genomics
R package version 4.4.0
\href{https://cran.r-project.org/web/packages/Seurat/}{Link}.
}
