% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoexpressionNetwork.R
\name{plotCoexpressionNetwork}
\alias{plotCoexpressionNetwork}
\title{Create a co-expression network graph. Each gene will be a node, and if the
genes are co-expressed, there will be an edge connecting them.}
\usage{
plotCoexpressionNetwork(edge_list)
}
\arguments{
\item{edge_list}{A list of all the genes that should have an edge between
them because they are co-expressed}
}
\value{
A network graph where co-expressed genes are connected by an edge
}
\description{
Create a co-expression network graph. Each gene will be a node, and if the
genes are co-expressed, there will be an edge connecting them.
}
\examples{
# Using pbmc_srat.rds file in data folder
srat <- readRds(file = "data/pbmc_srat.rds")
expr_matrix <- getExpressionMatrix(srat = srat)
corr_matrix <- getCorrelationMatrix(expression_matrix = expr_matrix)
edge_list <- buildCoexpressionNetwork(correlation_matrix = corr_matrix)
network_graph <- plotCoexpressionNetwork(edge_list)
network_graph

}
