% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelationMatrix.R
\name{plotCorrelationMatrix}
\alias{plotCorrelationMatrix}
\title{Plot a correlation matrix (columns and rows are genes and each cell contains
the correlation between the two genes)}
\usage{
plotCorrelationMatrix(correlation_matrix)
}
\arguments{
\item{correlation_matrix}{A matrix containing the correlation between all
pairs of genes}
}
\value{
A heatmap where each row and column is a gene. If the genes are
highly correlated (i.e. co-expressed), the cell will be close to red. If
the genes are not correlated (i.e. not co-expresseed), the cell will be
close to blue.
}
\description{
Plot a correlation matrix (columns and rows are genes and each cell contains
the correlation between the two genes)
}
