% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpressionMatrix.R
\name{plotExpressionMatrix}
\alias{plotExpressionMatrix}
\title{Plot an expression matrix (columns are cells and rows are genes)}
\usage{
plotExpressionMatrix(expression_matrix)
}
\arguments{
\item{expression_matrix}{A matrix containing gene counts per cell. Each
row is a gene and each column is a cell}
}
\value{
A heatmap where each row is a gene and each column is a cell. If the
gene is highly expressed in the cell, the cell will be close to red. If
the gene is lowly expressed in the cell, the cell will be close to blue.
}
\description{
Plot an expression matrix (columns are cells and rows are genes)
}
