% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Prepare the scRNA-seq data for analysis by normalizing, clustering, and
assigning cell type to cluster if applicable}
\usage{
prepareData(gene_bc_matrix_path, cell_types_path = NULL)
}
\arguments{
\item{gene_bc_matrix_path}{A path to filtered gene barcode matrices produced
using the cellranger pipeline.}

\item{cell_types_path}{A path to a csv file that maps marker genes to cell
type; default value is NULL (the clusters are not assigned a cell type).}
}
\value{
Returns an object of class SeuratObject that stores the data for
a single cell RNA-seq experiment
}
\description{
Prepare the scRNA-seq data for analysis by normalizing, clustering, and
assigning cell type to cluster if applicable
}
\examples{
# Using raw pbmc dataset available with package
# TODO: use system.file
gene_bc_matrix <- file.path('inst/extdata/pbmc/filtered_gene_bc_matrices')
cell_types <- file.path('inst/extdata/pbmc/cell_types.csv')

pbmc <- prepare_data(gene_bc_matrix_path = gene_bc_matrix,
                     cell_types_path = cell_types)
pbmc

}
\references{
Butler A, Hoffman P, Smibert P, Papalexi E, Satija R (2018).
“Integrating single-cell transcriptomic data across different conditions,
technologies, and species.” Nature Biotechnology, 36, 411-420.
\href{https://doi.org/10.1038/nbt.4096}{Link}.

Butler, A. (2015). Seurat: Tools for Single Cell Genomics
R package version 4.4.0
data \href{https://cran.r-project.org/web/packages/Seurat/}{Link}.
}
