% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoexpressionNetwork.R
\name{printCommunities}
\alias{printCommunities}
\title{Print what genes are part of the same community}
\usage{
printCommunities(networkGraph, communities)
}
\arguments{
\item{networkGraph}{an igraph object that contains some information about
a coexpression network}

\item{communities}{a list of communities mapping community names to the
genes that are in that community}
}
\description{
Print what genes are part of the same community
}
