% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDiffCoExpr.R
\name{runDiffCoExpr}
\alias{runDiffCoExpr}
\title{Launch Shiny App for DiffCoExpr}
\usage{
runDiffCoExpr()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for DiffCoExpr. The purpose of this
app is only to demonstrate how DiffCoExpr can be used to generate and
compare coexpression networks.
}
\details{
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

DiffCoExpr::runDiffCoExpr()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials.
\href{https://shiny.rstudio.com/tutorial/}{Link}
}
