% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrelationMatrix.R
\name{getCorrelationMatrix}
\alias{getCorrelationMatrix}
\title{Create (and visualize) the correlation between every pair of genes in the
expression matrix}
\usage{
getCorrelationMatrix(expression_matrix, visualize = FALSE)
}
\arguments{
\item{expression_matrix}{A matrix containing the expression levels of genes
per cell. The rows are genes and the columns are cells}

\item{visualize}{A boolean that will be used to determine if the function
should also create a heatmap of the correlation matrix. Default is set
to FALSE.}
}
\value{
A matrix where the rows and columns are genes, and the cell contains
the correlation between the two genes
}
\description{
Create (and visualize) the correlation between every pair of genes in the
expression matrix
}
\examples{
# Using saved Seurat object in the Data folder
srat <- readRds(file = "data/pbmc_srat.rds")
expr_matrix <- getExpressionMatrix(srat = srat)
corr_matrix <- getCorrelationMatrix(expression_matrix = expr_matrix)
corr_matrix

}
\references{
Lemoine G., Scott-Boyer M., Ambroise B., Perin O., Droit A. (2021) GWENA:
gene co-expression networks analysis and extended modules characterization
in a single Bioconductor package. BMC Bioinformatics 22, 267.
\href{https://doi.org/10.1186/s12859-021-04179-4}{Link}.
}
