% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpressionMatrix.R
\name{getExpressionMatrix}
\alias{getExpressionMatrix}
\title{Create an expression matrix for single-cell RNA-seq data. Each row is a
gene and each column is a cell.}
\usage{
getExpressionMatrix(srat, cellType = NULL, minCellCount = 5)
}
\arguments{
\item{srat}{A Seurat object containing normalized counts from a scRNA-seq
experiment.}

\item{cellType}{A cell type that should be used to filter the expression
matrix so all cells are of that cell type. Can also be a default cluster
name (e.g. c0, c1,...) if cell types were not used. Default is set to
NULL, so all cells of all types are included in the expression matrix.}

\item{minCellCount}{A minimum number of cells that should still be present
in the data after filtering. If the number of cells is less than this
value, stop running. Default is set to 5.}
}
\value{
A dense matrix where the rows are genes and the columns are cell
barcodes. Each cell contains the log normalized counts of that gene in
that cell.
}
\description{
Create an expression matrix for single-cell RNA-seq data. Each row is a
gene and each column is a cell.
}
\examples{
# Using pbmc dataset available with package
geneMatrixPath <- system.file("extdata",
                              "filtered_gene_bc_matrices",
                               package =  "DiffCoExpr")
cellTypes <- system.file("extdata", 
                         "cellTypes.csv", 
                          package = "DiffCoExpr")

pbmc <- prepareData(geneMatrixPath = geneMatrixPath,
                    cellTypesPath = cellTypes)
                    
exprMatrix <- getExpressionMatrix(srat = pbmc, 
                                  cellType = "Platelet")
exprMatrix

}
\references{
Butler, A. (2015). Seurat: Tools for Single Cell Genomics
R package version 4.4.0
\href{https://cran.r-project.org/web/packages/Seurat/}{Link}.
}
