% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoexpressionNetwork.R
\name{plotCoexpressionNetwork}
\alias{plotCoexpressionNetwork}
\title{Create a co-expression network graph. Each gene will be a node, and if the
genes are co-expressed, there will be an edge connecting them. If separate
communities/clusters of co-expressed genes are detected, each community
will be a different color. If there is not enough data to detect communities
accurately, communities will not be included, and instead the top 1000
genes (sorted by correlation) will be plotted.}
\usage{
plotCoexpressionNetwork(edgeList)
}
\arguments{
\item{edgeList}{A list of all the genes that should have an edge between
them because they are co-expressed}
}
\value{
A network graph where co-expressed genes are connected by an edge
}
\description{
Create a co-expression network graph. Each gene will be a node, and if the
genes are co-expressed, there will be an edge connecting them. If separate
communities/clusters of co-expressed genes are detected, each community
will be a different color. If there is not enough data to detect communities
accurately, communities will not be included, and instead the top 1000
genes (sorted by correlation) will be plotted.
}
\examples{
# Using an example coexpression network that was generated using the pbmc 
# data that is available in this package. This matrix only includes platelet
# cells.
coexprNetPath <- system.file("extdata", 
                             "coexpressionNetworkPlatelet.csv", 
                             package = "DiffCoExpr")
coexprNet <- read.csv(coexprNetPath)
# Delete the first column 
coexprNet <- coexprNet[, -1]

coexprNetworkGraph <- plotCoexpressionNetwork(edgeList = coexprNet)
coexprNetworkGraph

}
