% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDifferentialCoexpression.R
\name{plotDifferentialCoexpression}
\alias{plotDifferentialCoexpression}
\title{Visualize the distribution of differential co-expression that arise from
differences in co-expression levels between two conditions}
\usage{
plotDifferentialCoexpression(network_A, network_B)
}
\arguments{
\item{network_A}{An edge list for a co-expression network. Contains the
correlation coefficient between pairs of genes}

\item{network_B}{An edge list for a different co-expression network.
Contains the correlation coefficient between pairs of genes}
}
\value{
A histogram that illustrates the distribution of log fold changes.
The fold change refers to the degree to which co-expression strength
(e.g., correlation coefficient) changes between gene pairs across
co-expression networks constructed at different conditions (e.g. cell
types)
}
\description{
Visualize the distribution of differential co-expression that arise from
differences in co-expression levels between two conditions
}
