% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Prepare the scRNA-seq data for analysis by normalizing, clustering, and
assigning cell type to cluster if applicable}
\usage{
prepareData(geneMatrixPath, cellTypesPath = NULL)
}
\arguments{
\item{geneMatrixPath}{A path to the directory containing filtered gene
barcode matrices produced using the cellranger pipeline (as a reminder,
cellranger will produce a directory containing 3 files: matrix.mtx,
genes.tsv/features.tsv, and barcodes.tsv)}

\item{cellTypesPath}{A path to a csv file that maps marker genes to cell
type; default value is NULL (the clusters are not assigned a cell type).}
}
\value{
Returns an object of class SeuratObject that stores the data for
a single cell RNA-seq experiment
}
\description{
Prepare the scRNA-seq data for analysis by normalizing, clustering, and
assigning cell type to cluster if applicable
}
\examples{
# Using pbmc dataset available with package
geneMatrixPath <- system.file("extdata",
                              "filtered_gene_bc_matrices",
                               package =  "DiffCoExpr")
cellTypes <- system.file("extdata", 
                         "cellTypes.csv", 
                          package = "DiffCoExpr")

pbmc <- prepareData(geneMatrixPath = geneMatrixPath,
                    cellTypesPath = cellTypes)
pbmc

}
\references{
Butler A, Hoffman P, Smibert P, Papalexi E, Satija R (2018).
“Integrating single-cell transcriptomic data across different conditions,
technologies, and species.” Nature Biotechnology, 36, 411-420.
\href{https://doi.org/10.1038/nbt.4096}{Link}.

Butler, A. (2015). Seurat: Tools for Single Cell Genomics
R package version 4.4.0
data \href{https://cran.r-project.org/web/packages/Seurat/}{Link}.
}
